Text Moved to Calendar Component
