+++ src/components/Calendar/index.tsx
@@ -35,7 +35,8 @@ export default function Calendar({
   capture = false,
   isAutosized = false,
 }: CalendarProps): React.ReactElement {
-  const [{ pinnedCrns, oscar, events }] = useContext(ScheduleContext);
+  const [{ pinnedCrns, oscar, events, desiredCourses }] =
+    useContext(ScheduleContext);
 
   // Contains the rowIndex's and rowSize's passed into each crn's TimeBlocks
   // e.g. crnSizeInfo[crn][day]["period.start-period.end"].rowIndex
@@ -233,6 +234,29 @@ export default function Calendar({
     }
   });
 
+  // const [{ oscar, desiredCourses, pinnedCrns }] = useContext(ScheduleContext);
+
+  const filteredCourses = oscar.courses.filter((course) => {
+    if (desiredCourses.includes(course.id)) {
+      return course;
+    }
+    return '';
+  });
+
+  const finalFilteredCourses = filteredCourses.filter((course) => {
+    let match;
+    course.sections.forEach((section) => {
+      if (
+        section.campus !== 'Georgia Tech-Atlanta *' ||
+        section.meetings[0]?.days.includes('S')
+      ) {
+        match = course;
+      }
+      return '';
+    });
+    return match;
+  });
+
   return (
     <div
       className={classes(
@@ -327,6 +351,26 @@ export default function Calendar({
             />
           ))}
       </div>
+      {!preview ? (
+        <div className="hidden-courses">
+          *Other Courses/Events not shown in view:{' '}
+          {finalFilteredCourses.map((course) => {
+            let sectionId = '';
+            pinnedCrns.filter((crn) => {
+              course.sections.every((section) => {
+                if (section.crn === crn) {
+                  sectionId = section.id;
+                }
+                return 'i';
+              });
+              return 'j';
+            });
+            return `${course.id}(${sectionId}), `;
+          })}
+        </div>
+      ) : (
+        <div />
+      )}
     </div>
   );
 }
